refac: BIG ASS REFAC
- Removed CRUD methods from EncounterService and inherited CrudService
- Encounter module now uses Location from BuildingBlocks
- Updated ICrudService to only require TDto

+++ src/Explorer.API/Controllers/Administrator/Administration/EncounterController.cs
@@ -41,7 +41,7 @@ public class EncounterController : BaseApiController
     [HttpDelete("{id:long}")]
     public ActionResult Delete(long id)
     {
-        var result = _encounterService.Delete(id);
+        var result = _encounterService.Delete((int) id);
         return CreateResponse(result);
     }
 }
